% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarPlotETL.R
\name{artists_popularity}
\alias{artists_popularity}
\title{Get Top Twenty Artists Ranked by Popularity}
\usage{
artists_popularity(artist_id)
}
\arguments{
\item{artist_id}{Character string of the search artist's Spotify Id}
}
\value{
bar plot
}
\description{
This function returns an interactive bar plot of the top twenty most popular
artists in the network.
}
\examples{
artists_popularity(artist_id = "3fMbdgg4jU18AjLCKBhRSm")
}
