% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeETL.R
\name{related_artists_edges}
\alias{related_artists_edges}
\title{Get Network Edges Data}
\usage{
related_artists_edges(artist_id)
}
\arguments{
\item{artist_id}{Character string of the search artist's Spotify Id}
}
\value{
table
}
\description{
This function returns an interactive table of the edges in
the network graph.
}
\examples{
related_artists_edges(artist_id = "3fMbdgg4jU18AjLCKBhRSm")
}
