% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkETL.R
\name{related_artists_network}
\alias{related_artists_network}
\title{Get Related Artists Network Graph}
\usage{
related_artists_network(artist_id)
}
\arguments{
\item{artist_id}{Character string of the search artist's Spotify Id}
}
\value{
network graph
}
\description{
This function returns an interactive network graph of artists
that are related to the search artist.
}
\examples{
related_artists_network(artist_id = "3fMbdgg4jU18AjLCKBhRSm")
}
