% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeETL.R
\name{related_artists_nodes}
\alias{related_artists_nodes}
\title{Get Network Nodes Data}
\usage{
related_artists_nodes(artist_id)
}
\arguments{
\item{artist_id}{Character string of the search artist's Spotify Id}
}
\value{
table
}
\description{
This function returns an interactive table of the artists in the
network and their features.
}
\examples{
related_artists_nodes(artist_id = "3fMbdgg4jU18AjLCKBhRSm")
}
